add sample effect and bit crusher init

+++ components/mixer/include/mixer.h
@@ -59,7 +59,7 @@ typedef struct sample_t
 {
     const unsigned char *raw_data; /** raw sample bytes */
     wav_header_t header; /** contains sample metadata like size and bit rate */
-    uint32_t playback_ptr; /** progress indicator for the sample */
+    float playback_ptr; /** progress indicator for the sample */
     // playback_mode_t playback_mode; /** sample play type: ONESHOT, LOOP, etc... */
     int sample_id;
     bool playback_finished;
